refactor(emitter): split IR printer recovery helpers
